solved rounding error in dry/wet but dry/wet values still dont seem to influence the sound alot

+++ CSD2c/final_assignment/code/Callback.h
@@ -16,7 +16,7 @@ public:
   void setOsc(float oscValue);
 
 private:
-  int oscValue = 0;
+  float oscValue = 0.f;
   EffectController effects;
   float samplerate = 44100;
   int frameIndex = 0;
